refactor: separate error handlers of app in middleware dir
